fix: resolve 31 maintainability violations across 31 files
- Replace all wildcard imports in 11 shim files with explicit named imports
- Remove private symbol re-exports from 8 backward-compat shims
- Split mappers.py (456 lines) into 4 sub-modules, each under 300 lines
- Split jobs.py Job/JobStore model into _job_model.py
- Extract _start_action_api helper in dashboard runner (reduce callees)
- Group _spawn_and_monitor params into _SpawnPaths dataclass
- Make _PERCENT_SCALE public (PERCENT_SCALE) in evidence model
- Rename _read_message to read_message in MCP dispatch
- Use _MAX_PENALTY_MULTIPLIER constant instead of literal 1.30
- Add injectable env param to cli.py _env_int and _subagent_model
- Add create_accumulated_cache() factory for swappable cache
- Extract I/O and security helpers from utils.py into _io.py and _security.py
