+++ src/quodeq/engine/evidence_parser.py
@@ -1,3 +1,8 @@
 """Re-export for backward compatibility — moved to quodeq.core.evidence.parser."""
-from quodeq.core.evidence.parser import *  # noqa: F401,F403
+from quodeq.core.evidence.parser import (  # noqa: F401
+    build_req_refs_lookup,
+    EvidenceContext,
+    resolve_llm_refs,
+    parse_jsonl_to_evidence,
+)
 from quodeq.core.evidence.parser import _parse_jsonl_line  # noqa: F401
