+++ src/quodeq/engine/report.py
@@ -6,7 +6,4 @@ from quodeq.analysis.report import (  # noqa: F401
     build_dashboard_report,
     write_reports,
     write_dimension_report,
-    _build_score_lookup,
-    _flatten_findings,
-    _build_principle_rows,
 )
