+++ src/quodeq/provider/_cache.py
@@ -1,2 +1,2 @@
 """Re-export for backward compatibility — moved to quodeq.services._cache."""
-from quodeq.services._cache import *  # noqa: F401,F403
+from quodeq.services._cache import make_lru_dimension_fetcher  # noqa: F401
