+++ src/quodeq/provider/_dashboard_stale.py
@@ -1,2 +1,2 @@
 """Re-export for backward compatibility — moved to quodeq.services._dashboard_stale."""
-from quodeq.services._dashboard_stale import *  # noqa: F401,F403
+from quodeq.services._dashboard_stale import collect_stale_dimensions  # noqa: F401
