+++ src/quodeq/provider/base.py
@@ -1,2 +1,9 @@
 """Re-export for backward compatibility — moved to quodeq.services.base."""
-from quodeq.services.base import *  # noqa: F401,F403
+from quodeq.services.base import (  # noqa: F401
+    EvaluationOptions,
+    ProjectActions,
+    ReportActions,
+    EvaluationActions,
+    ToolingActions,
+    ActionProvider,
+)
