+++ src/quodeq/provider/filesystem.py
@@ -1,2 +1,2 @@
 """Re-export for backward compatibility — moved to quodeq.services.filesystem."""
-from quodeq.services.filesystem import *  # noqa: F401,F403
+from quodeq.services.filesystem import FilesystemActionProvider  # noqa: F401
