+++ src/quodeq/provider/plugin_discovery.py
@@ -1,2 +1,2 @@
 """Re-export for backward compatibility — moved to quodeq.services.plugin_discovery."""
-from quodeq.services.plugin_discovery import *  # noqa: F401,F403
+from quodeq.services.plugin_discovery import discover_plugins  # noqa: F401
