+++ src/quodeq/provider/tooling_mixin.py
@@ -1,2 +1,5 @@
 """Re-export for backward compatibility — moved to quodeq.services.tooling_mixin."""
-from quodeq.services.tooling_mixin import *  # noqa: F401,F403
+from quodeq.services.tooling_mixin import (  # noqa: F401
+    get_allowed_client_ids,
+    FsToolingMixin,
+)
