+++ src/quodeq/provider/violations.py
@@ -1,2 +1,6 @@
 """Re-export for backward compatibility — moved to quodeq.services.violations."""
-from quodeq.services.violations import *  # noqa: F401,F403
+from quodeq.services.violations import (  # noqa: F401
+    resolve_dimension_eval,
+    aggregate_violations,
+    ViolationContext,
+)
