+++ tests/test_action_provider_fs_dashboard.py
@@ -8,7 +8,7 @@ from typing import Any
 import pytest
 
 from quodeq.core.types import DimensionResult
-from quodeq.provider.dashboard import (
+from quodeq.services.dashboard import (
     _collect_previous_scores,
     _collect_stale_dimensions,
     _enrich_dimensions_with_trend,
